requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
